% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgarten_packages.R
\name{statgarten_packages}
\alias{statgarten_packages}
\title{List all packages in the statgarten}
\usage{
statgarten_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include statgarten in the list?}
}
\description{
List all packages in the statgarten
}
\examples{
statgarten_packages()
}
